backend/python/requirements.txt
google-generativeai==0.3.2
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
absl-py==1.4.0 